Reached the limit of Rust's type system
Rust's type system does not allow specialization (in Stable Rust), and
even the Nightly implementations are quite limiting and buggy.
Additionally, it is apparently not possible to implement `std::ops::Add`
for both `T` and `&T`, since `&T` is a subset of `T`. It *should* be
possible, but for some reason private traits are still covered by Rust's
Orphan Rules, so I can't get around it that way. Plz send help :(